services/judge/pyproject.toml
[project]
name = "botcheck-judge"
version = "0.1.0"
description = "BotCheck Judge Service — deterministic + LLM scoring of test runs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "arq>=0.26",
    "anthropic>=0.40",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "aioboto3>=13.0",
    "prometheus-client>=0.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
    "opentelemetry-instrumentation-httpx>=0.40b0",
    "opentelemetry-instrumentation-anthropic>=0.52",
    "botcheck-observability",
    "botcheck-scenarios",
]

[tool.uv.sources]
botcheck-observability = { workspace = true }
botcheck-scenarios = { workspace = true }

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["botcheck_judge"]
